Refactor the structure of the project
