+++ lookout/core/analyzer.py
@@ -6,6 +6,48 @@ from lookout.core.api.service_analyzer_pb2 import Comment
 from lookout.core.api.service_data_pb2_grpc import DataStub
 
 
+class AnalyzerModel(Model):
+    """
+    All models used in `Analyzer`-s must derive from this base class.
+    """
+    def __init__(self, **kwargs):
+        """
+        Defines:
+        `name` - name of the model. Corresponds to the bound analyzer's class name and version.
+        `url` - Git repository on which the model was trained.
+        `commit` - revision of the Git repository on which the model was trained.
+        :param kwargs: passed to the upstream __init__.
+        """
+        super().__init__(**kwargs)
+        self.name = "<unknown name>"
+        self.url = "<unknown url>"
+        self.commit = "<unknown commit>"
+
+    def construct(self, analyzer: Type["Analyzer"], url: str, commit: str):
+        """
+        Initializing method.
+
+        :param analyzer: Bound type of the `Analyzer`. Not instance!
+        :param url: Git repository on which the model was trained.
+        :param commit: revision of the Git repository on which the model was trained.
+        :return: self
+        """
+        assert isinstance(self, analyzer.model_type)
+        self.name = analyzer.__name__
+        self.version = [analyzer.version]
+        self.url = url
+        self.commit = commit
+        return self
+
+    def dump(self) -> str:
+        """
+        Implements the upstream abstract method.
+
+        :return: summary text of the model.
+        """
+        return "%s/%s %s %s" % (self.name, self.version, self.url, self.commit)
+
+
 class Analyzer:
     """
     Interface of all the analyzers. Each analyzer uses a model to run the analysis and generates
@@ -17,9 +59,9 @@ class Analyzer:
     in analyze() and generated in train().
     """
     version = None  # type: str
-    model_type = None  # type: Type[Model]
+    model_type = None  # type: Type[AnalyzerModel]
 
-    def __init__(self, model: Model, url: str, config: dict):
+    def __init__(self, model: AnalyzerModel, url: str, config: dict):
         """
         :param model: The instance of the model loaded from the repository or freshly trained.
         :param url: The analyzed project's Git remote.
@@ -49,7 +91,7 @@ class Analyzer:
 
     @classmethod
     def train(cls, url: str, commit: str, config: dict, data_request_stub: DataStub,
-              **data) -> Model:
+              **data) -> AnalyzerModel:
         """
         Generates a new model on top of the specified source code.
