+++ lookout/core/tests/test_manager.py
@@ -1,0 +1,9 @@
+import unittest
+
+
+class AnalyzerManagerTests(unittest.TestCase):
+    pass
+
+
+if __name__ == "__main__":
+    unittest.main()
