+++ requirements.txt
@@ -10,3 +10,4 @@ humanfriendly==4.16.1
 psycopg2-binary==2.7.5
 scikit-learn==0.19.2
 tqdm==4.11.2
+modelforge==0.6.4
